+++ src/Fenix.Runtime/Common/Message/RegisterReq.cs
@@ -14,7 +14,7 @@ namespace Fenix.Common.Message
     public class RegisterReq : IMessage
     {
         [Key(0)]
-        public UInt32 hostId { get; set; }
+        public UInt64 hostId { get; set; }
 
         [Key(1)]
         public String hostName { get; set; }
